MULE-18029: Changes due to SuppressedMuleException moving to internal
